fix(graph): split store.rs tests into their own module to clear the size guard
`main` is red on `check-file-size`. `stella-graph/src/store.rs` reached 1632
lines when #1313 (`4a0323f6`) landed the call-site recording work, crossing the
1500-line limit.

This is the guard's *non*-grandfathered branch, which is a different failure
from the baseline drift #1310 just repaired. The message is explicit that a
baseline entry is not the remedy here:

  These files crossed the limit and are NOT grandfathered — split them into
  submodules. Do not add a baseline entry: the baseline only covers files
  that predate the guard, and this is the rule that stops the tree getting
  worse.

So the inline `#[cfg(test)] mod tests` moves to `stella-graph/src/store/tests.rs`
and `store.rs` declares it as a file module — the same shape `command_deck.rs`,
`driver.rs` and `pipeline.rs` already use. Production code lands at 1060 lines
and the tests at 571, both under the limit with room to spare, and neither file
needs a baseline row.

The move is mechanical: the module path is unchanged, so all 18 tests still run
as `store::tests::*` — none renamed, none dropped. `cargo test -p stella-graph`
is 112 passed / 0 failed. Indentation was normalised with rustfmt rather than a
blind dedent, so no string literal contents shifted.

Verified on this branch by exit code: check-file-size 0, cargo fmt --all --check
0, shellcheck 0, cargo clippy --workspace --all-targets -D warnings 0.

Note for review: #1313 merged while its own `fmt + clippy + test` was still
pending, which is why this reached `main` at all.
